vendor: dts: usb0: force peripheral mode
Aeroratech CCB board has a USB switch connects to USB0 port
with 2 TypeC outputs. VBUS is always ON to USB0 from the USB
switch so there will be no VBUS trigger and notification in
kernel driver for proper mode change/settings.

For usb0 to work properly in device mode require to force USB0
to stay in peripheral mode.